+++ code/qcommon/parse.c
@@ -1,7 +1,7 @@
 /*
 ===========================================================================
 Copyright (C) 1999-2005 Id Software, Inc.
-Copyright (C) 2000-2006 Tim Angus
+Copyright (C) 2000-2013 Darklegion Development
 
 This file is part of Tremulous.
 
@@ -127,7 +127,7 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 //punctuation
 typedef struct punctuation_s
 {
-  const qchar *p;                    //punctuation character(s)
+  qchar *p;                    //punctuation character(s)
   qint n;                      //punctuation indication
   struct punctuation_s *next; //next punctuation
 } punctuation_t;
@@ -234,6 +234,8 @@ typedef struct directive_s
 #define DEFINEHASHSIZE    1024
 
 static qint Parse_ReadToken(source_t *source, token_t *token);
+static qbool Parse_AddDefineToSourceFromString( source_t *source,
+                                                   const qchar *string );
 
 qint numtokens;
 
@@ -257,7 +259,7 @@ punctuation_t default_punctuations[] =
   {"<=",P_LOGIC_LEQ, NULL},
   {"==",P_LOGIC_EQ, NULL},
   {"!=",P_LOGIC_UNEQ, NULL},
-  //arithmetic operators
+  //arithmatic operators
   {"*=",P_MUL_ASSIGN, NULL},
   {"/=",P_DIV_ASSIGN, NULL},
   {"%=",P_MOD_ASSIGN, NULL},
@@ -276,7 +278,7 @@ punctuation_t default_punctuations[] =
   //C++
   {"::",P_CPP1, NULL},
   {".*",P_CPP2, NULL},
-  //arithmetic operators
+  //arithmatic operators
   {"*",P_MUL, NULL},
   {"/",P_DIV, NULL},
   {"%",P_MOD, NULL},
@@ -294,7 +296,7 @@ punctuation_t default_punctuations[] =
   {"<",P_LOGIC_LESS, NULL},
   //reference operator
   {".",P_REF, NULL},
-  //separators
+  //seperators
   {",",P_COMMA, NULL},
   {";",P_SEMICOLON, NULL},
   //label indication
@@ -315,8 +317,6 @@ punctuation_t default_punctuations[] =
   {"$",P_DOLLAR, NULL},
   {NULL, 0}
 };
-
-qchar basefolder[MAX_QPATH];
 
 /*
 ===============
@@ -363,15 +363,15 @@ static void Parse_CreatePunctuationTable(script_t *script, punctuation_t *punctu
 Parse_ScriptError
 ===============
 */
-static void QDECL Parse_ScriptError(script_t *script, const qchar *fmt, ...)
+__attribute__ ((format (printf, 2, 3))) static void QDECL Parse_ScriptError(script_t *script, qchar *str, ...)
 {
   qchar text[1024];
   va_list ap;
 
   if (script->flags & SCFL_NOERRORS) return;
 
-  va_start(ap, fmt);
-  Q_vsnprintf(text, sizeof(text), fmt, ap);
+  va_start(ap, str);
+  vsprintf(text, str, ap);
   va_end(ap);
   Com_Printf( "file %s, line %d: %s\n", script->filename, script->line, text);
 }
@@ -381,15 +381,15 @@ static void QDECL Parse_ScriptError(script_t *script, const qchar *fmt, ...)
 Parse_ScriptWarning
 ===============
 */
-static void QDECL Parse_ScriptWarning(script_t *script, const qchar *fmt, ...)
+__attribute__ ((format (printf, 2, 3))) static void QDECL Parse_ScriptWarning(script_t *script, qchar *str, ...)
 {
   qchar text[1024];
   va_list ap;
 
   if (script->flags & SCFL_NOWARNINGS) return;
 
-  va_start(ap, fmt);
-  Q_vsnprintf(text, sizeof(text), fmt, ap);
+  va_start(ap, str);
+  vsprintf(text, str, ap);
   va_end(ap);
   Com_Printf( "file %s, line %d: %s\n", script->filename, script->line, text);
 }
@@ -588,7 +588,7 @@ static qint Parse_ReadString(script_t *script, token_t *token, qint quote)
       //
       tmpscript_p = script->script_p;
       tmpline = script->line;
-      //read unuseful stuff between possible two following strings
+      //read unusefull stuff between possible two following strings
       if (!Parse_ReadWhiteSpace(script))
       {
         script->script_p = tmpscript_p;
@@ -770,7 +770,7 @@ static qint Parse_ReadNumber(script_t *script, token_t *token)
     }
     token->subtype |= TT_HEX;
   }
-#if defined(BINARYNUMBERS)
+#ifdef BINARYNUMBERS
   //check for a binary number
   else if (*script->script_p == '0' &&
     (*(script->script_p + 1) == 'b' ||
@@ -847,7 +847,7 @@ Parse_ReadPunctuation
 static qint Parse_ReadPunctuation(script_t *script, token_t *token)
 {
   qint len;
-  const qchar *p;
+  qchar *p;
   punctuation_t *punc;
 
   for (punc = script->punctuationtable[(unsigned qint)*script->script_p]; punc; punc = punc->next)
@@ -860,7 +860,7 @@ static qint Parse_ReadPunctuation(script_t *script, token_t *token)
       //if the script contains the punctuation
       if (!strncmp(script->script_p, p, len))
       {
-        Q_strncpyz(token->string, p, sizeof(token->string));
+        strncpy(token->string, p, MAX_TOKEN_CHARS);
         script->script_p += len;
         token->type = TT_PUNCTUATION;
         //sub type is the number of the punctuation
@@ -884,7 +884,7 @@ static qint Parse_ReadPrimitive(script_t *script, token_t *token)
   len = 0;
   while(*script->script_p > ' ' && *script->script_p != ';')
   {
-    if (len >= MAX_TOKEN_CHARS - 1)
+    if (len >= MAX_TOKEN_CHARS)
     {
       Parse_ScriptError(script, "primitive token longer than MAX_TOKEN_CHARS = %d", MAX_TOKEN_CHARS);
       return 0;
@@ -894,7 +894,7 @@ static qint Parse_ReadPrimitive(script_t *script, token_t *token)
   token->string[len] = 0;
   //copy the token into the script structure
   Com_Memcpy(&script->token, token, sizeof(token_t));
-  //primitive reading successful
+  //primitive reading successfull
   return 1;
 }
 
@@ -921,7 +921,7 @@ static qint Parse_ReadScriptToken(script_t *script, token_t *token)
   //start of the white space
   script->whitespace_p = script->script_p;
   token->whitespace_p = script->script_p;
-  //read unuseful stuff
+  //read unusefull stuff
   if (!Parse_ReadWhiteSpace(script)) return 0;
 
   script->endwhitespace_p = script->script_p;
@@ -981,7 +981,7 @@ static void Parse_StripDoubleQuotes(qchar *string)
 {
   if (*string == '\"')
   {
-    memmove(string, string + 1, strlen(string));
+    memmove( string, string + 1, strlen( string ) + 1 );
   }
   if (string[strlen(string)-1] == '\"')
   {
@@ -1007,16 +1007,11 @@ Parse_LoadScriptFile
 static script_t *Parse_LoadScriptFile(const qchar *filename)
 {
   fileHandle_t fp;
-  qchar pathname[MAX_QPATH];
   qint length;
   void *buffer;
   script_t *script;
 
-  if (strlen(basefolder))
-    Com_sprintf(pathname, sizeof(pathname), "%s/%s", basefolder, filename);
-  else
-    Com_sprintf(pathname, sizeof(pathname), "%s", filename);
-  length = FS_FOpenFileRead( pathname, &fp, qfalse );
+  length = FS_FOpenFileRead( filename, &fp, qfalse );
   if (!fp) return NULL;
 
   buffer = Z_Malloc(sizeof(script_t) + length + 1);
@@ -1024,7 +1019,7 @@ static script_t *Parse_LoadScriptFile(const qchar *filename)
 
   script = (script_t *) buffer;
   Com_Memset(script, 0, sizeof(script_t));
-  Q_strncpyz(script->filename, filename, sizeof(script->filename));
+  strcpy(script->filename, filename);
   script->buffer = (qchar *) buffer + sizeof(script_t);
   script->buffer[length] = 0;
   script->length = length;
@@ -1054,7 +1049,7 @@ static script_t *Parse_LoadScriptFile(const qchar *filename)
 Parse_LoadScriptMemory
 ===============
 */
-static script_t *Parse_LoadScriptMemory(const qchar *ptr, qint length, const qchar *name)
+static script_t *Parse_LoadScriptMemory(const qchar *ptr, qint length, qchar *name)
 {
   void *buffer;
   script_t *script;
@@ -1064,7 +1059,7 @@ static script_t *Parse_LoadScriptMemory(const qchar *ptr, qint length, const qch
 
   script = (script_t *) buffer;
   Com_Memset(script, 0, sizeof(script_t));
-  Q_strncpyz(script->filename, name, sizeof(script->filename));
+  strcpy(script->filename, name);
   script->buffer = (qchar *) buffer + sizeof(script_t);
   script->buffer[length] = 0;
   script->length = length;
@@ -1100,26 +1095,16 @@ static void Parse_FreeScript(script_t *script)
 
 /*
 ===============
-Parse_SetBaseFolder
-===============
-*/
-static void Parse_SetBaseFolder(const qchar *path)
-{
-  Com_sprintf(basefolder, sizeof(basefolder), "%s", path);
-}
-
-/*
-===============
 Parse_SourceError
 ===============
 */
-static void QDECL Parse_SourceError(source_t *source, const qchar *fmt, ...)
+__attribute__ ((format (printf, 2, 3))) static void QDECL Parse_SourceError(source_t *source, qchar *str, ...)
 {
   qchar text[1024];
   va_list ap;
 
-  va_start(ap, fmt);
-  Q_vsnprintf(text, sizeof(text), fmt, ap);
+  va_start(ap, str);
+  vsprintf(text, str, ap);
   va_end(ap);
   Com_Printf( "file %s, line %d: %s\n", source->scriptstack->filename, source->scriptstack->line, text);
 }
@@ -1129,13 +1114,13 @@ static void QDECL Parse_SourceError(source_t *source, const qchar *fmt, ...)
 Parse_SourceWarning
 ===============
 */
-static void QDECL Parse_SourceWarning(source_t *source, const qchar *fmt, ...)
+__attribute__ ((format (printf, 2, 3))) static void QDECL Parse_SourceWarning(source_t *source, qchar *str, ...)
 {
   qchar text[1024];
   va_list ap;
 
-  va_start(ap, fmt);
-  Q_vsnprintf(text, sizeof(text), fmt, ap);
+  va_start(ap, str);
+  vsprintf(text, str, ap);
   va_end(ap);
   Com_Printf( "file %s, line %d: %s\n", source->scriptstack->filename, source->scriptstack->line, text);
 }
@@ -1149,7 +1134,7 @@ static void Parse_PushIndent(source_t *source, qint type, qint skip)
 {
   indent_t *indent;
 
-  indent = (indent_t *)Z_Malloc(sizeof(indent_t));
+  indent = (indent_t *) Z_Malloc(sizeof(indent_t));
   indent->type = type;
   indent->script = source->scriptstack;
   indent->skip = (skip != 0);
@@ -1219,7 +1204,7 @@ static token_t *Parse_CopyToken(token_t *token)
 //  t = freetokens;
   if (!t)
   {
-    Com_Error(ERR_FATAL, "out of token space");
+    Com_Error(ERR_FATAL, "out of token space\n");
     return NULL;
   }
 //  freetokens = freetokens->next;
@@ -1252,13 +1237,22 @@ static qint Parse_ReadSourceToken(source_t *source, token_t *token)
 {
   token_t *t;
   script_t *script;
-  qint type, skip;
+  qint type, skip, lines;
 
+  lines = 0;
   //if there's no token already available
   while(!source->tokens)
   {
     //if there's a token to read from the script
-    if (Parse_ReadScriptToken(source->scriptstack, token)) return qtrue;
+    if( Parse_ReadScriptToken( source->scriptstack, token ) )
+    {
+      token->linescrossed += lines;
+      return qtrue;
+    }
+
+    // if lines were crossed before the end of the script, count them
+    lines += source->scriptstack->line - source->scriptstack->lastline;
+
     //if at the end of the script
     if (Parse_EndOfScript(source->scriptstack))
     {
@@ -1361,6 +1355,7 @@ static qint Parse_ReadDefineParms(source_t *source, define_t *define, token_t **
         if (indent <= 0)
         {
           if (lastcomma) Parse_SourceWarning(source, "too many comma's");
+          lastcomma = 1;
           break;
         }
       }
@@ -1404,33 +1399,20 @@ static qint Parse_ReadDefineParms(source_t *source, define_t *define, token_t **
 Parse_StringizeTokens
 ===============
 */
-static qint Parse_StringizeTokens(const token_t *tokens, token_t *token)
+static qint Parse_StringizeTokens(token_t *tokens, token_t *token)
 {
-  const token_t *t;
-  qint len;
-  qint total;
+  token_t *t;
 
   token->type = TT_STRING;
   token->whitespace_p = NULL;
   token->endwhitespace_p = NULL;
-  token->string[0] = '"';
-  total = 1;
-
+  token->string[0] = '\0';
+  strcat(token->string, "\"");
   for (t = tokens; t; t = t->next)
   {
-    len = (qint)strlen(t->string);
-
-    if (len + total >= sizeof(token->string) - 1) //reserve space for '"' and '\0'
-    {
-      return qfalse;
-    }
-
-    strcpy(token->string + total, t->string);
-    total += len;
+    strncat(token->string, t->string, MAX_TOKEN_CHARS - strlen(token->string));
   }
-
-  strcpy(token->string + total, "\"");
-
+  strncat(token->string, "\"", MAX_TOKEN_CHARS - strlen(token->string));
   return qtrue;
 }
 
@@ -1444,27 +1426,14 @@ static qint Parse_MergeTokens(token_t *t1, token_t *t2)
   //merging of a name with a name or number
   if (t1->type == TT_NAME && (t2->type == TT_NAME || t2->type == TT_NUMBER))
   {
-    if (strlen(t1->string) + strlen(t2->string) >= sizeof(t1->string))
-    {
-      return qfalse;
-    }
-
     strcat(t1->string, t2->string);
     return qtrue;
   }
-
   //merging of two strings
   if (t1->type == TT_STRING && t2->type == TT_STRING)
   {
-    qint len1 = (qint)strlen( t1->string );
-
-    if (strlen(t1->string) + strlen(t2->string) - 2 >= sizeof(t1->string))
-    {
-      return qfalse;
-    }
-
     //remove trailing double quote
-    t1->string[len1 - 1] = '\0';
+    t1->string[strlen(t1->string)-1] = '\0';
     //concat without leading double quote
     strcat(t1->string, &t2->string[1]);
     return qtrue;
@@ -1481,7 +1450,7 @@ Parse_NameHash
 //qchar primes[16] = {1, 3, 5, 7, 11, 13, 17, 19, 23, 27, 29, 31, 37, 41, 43, 47};
 static qint Parse_NameHash(qchar *name)
 {
-  qint hash, i;
+  qint register hash, i;
 
   hash = 0;
   for (i = 0; name[i] != '\0'; i++)
@@ -1531,7 +1500,7 @@ static define_t *Parse_FindHashedDefine(define_t **definehash, qchar *name)
 Parse_FindDefineParm
 ===============
 */
-static qint Parse_FindDefineParm(define_t *define, const qchar *name)
+static qint Parse_FindDefineParm(define_t *define, qchar *name)
 {
   token_t *p;
   qint i;
@@ -1609,11 +1578,12 @@ static qint Parse_ExpandBuiltinDefine(source_t *source, token_t *deftoken, defin
     case BUILTIN_DATE:
     {
       t = time(NULL);
-      curtime = ctime((const time_t *)&t);
+      curtime = ctime(&t);
       strcpy(token->string, "\"");
       strncat(token->string, curtime+4, 7);
       strncat(token->string+7, curtime+20, 4);
       strcat(token->string, "\"");
+      free(curtime);
       token->type = TT_NAME;
       token->subtype = strlen(token->string);
       *firsttoken = token;
@@ -1623,10 +1593,11 @@ static qint Parse_ExpandBuiltinDefine(source_t *source, token_t *deftoken, defin
     case BUILTIN_TIME:
     {
       t = time(NULL);
-      curtime = ctime((const time_t *)&t);
+      curtime = ctime(&t);
       strcpy(token->string, "\"");
       strncat(token->string, curtime+11, 8);
       strcat(token->string, "\"");
+      free(curtime);
       token->type = TT_NAME;
       token->subtype = strlen(token->string);
       *firsttoken = token;
@@ -1652,7 +1623,7 @@ Parse_ExpandDefine
 static qint Parse_ExpandDefine(source_t *source, token_t *deftoken, define_t *define,
                     token_t **firsttoken, token_t **lasttoken)
 {
-  token_t *parms[MAX_DEFINEPARMS] = {NULL}, *dt, *pt, *t;
+  token_t *parms[MAX_DEFINEPARMS], *dt, *pt, *t;
   token_t *t1, *t2, *first, *last, *nextpt, token;
   qint parmnum, i;
 
@@ -1801,7 +1772,7 @@ static void Parse_ConvertPath(qchar *path)
 {
   qchar *ptr;
 
-  //remove double path separators
+  //remove double path seperators
   for (ptr = path; *ptr;)
   {
     if ((*ptr == '\\' || *ptr == '/') &&
@@ -1814,10 +1785,10 @@ static void Parse_ConvertPath(qchar *path)
       ptr++;
     }
   }
-  //set OS dependent path separators
+  //set OS dependent path seperators
   for (ptr = path; *ptr;)
   {
-    if (*ptr == '/' || *ptr == '\\') *ptr = PATHSEPERATOR_CHAR;
+    if (*ptr == '/' || *ptr == '\\') *ptr = PATH_SEP;
     ptr++;
   }
 }
@@ -1919,7 +1890,7 @@ static qint Parse_OperatorPriority(qint op)
 //
 #define AllocOperator(op)               \
   if (numoperators >= MAX_OPERATORS) {        \
-    Parse_SourceError(source, "out of operator space"); \
+    Parse_SourceError(source, "out of operator space\n"); \
     error = 1;                    \
     break;                      \
   }                         \
@@ -1932,7 +1903,8 @@ static qint Parse_OperatorPriority(qint op)
 Parse_EvaluateTokens
 ===============
 */
-static qint Parse_EvaluateTokens(source_t *source, token_t *tokens, qint *intvalue, float *floatvalue, qint integer)
+static qint Parse_EvaluateTokens(source_t *source, token_t *tokens, signed long qint *intvalue,
+                                  double *floatvalue, qint integer)
 {
   operator_t *o, *firstoperator, *lastoperator;
   value_t *v, *firstvalue, *lastvalue, *v1, *v2;
@@ -2081,7 +2053,7 @@ static qint Parse_EvaluateTokens(source_t *source, token_t *tokens, qint *intval
             t->subtype == P_BIN_AND || t->subtype == P_BIN_OR ||
             t->subtype == P_BIN_XOR)
           {
-            Parse_SourceError(source, "illigal operator %s on floating point operands", t->string);
+            Parse_SourceError(source, "illigal operator %s on floating point operands\n", t->string);
             error = 1;
             break;
           }
@@ -2236,7 +2208,7 @@ static qint Parse_EvaluateTokens(source_t *source, token_t *tokens, qint *intval
                   v1->floatvalue *= v2->floatvalue; break;
       case P_DIV:       if (!v2->intvalue || !v2->floatvalue)
                   {
-                    Parse_SourceError(source, "divide by zero in #if/#elif");
+                    Parse_SourceError(source, "divide by zero in #if/#elif\n");
                     error = 1;
                     break;
                   }
@@ -2244,7 +2216,7 @@ static qint Parse_EvaluateTokens(source_t *source, token_t *tokens, qint *intval
                   v1->floatvalue /= v2->floatvalue; break;
       case P_MOD:       if (!v2->intvalue)
                   {
-                    Parse_SourceError(source, "divide by zero in #if/#elif");
+                    Parse_SourceError(source, "divide by zero in #if/#elif\n");
                     error = 1;
                     break;
                   }
@@ -2320,12 +2292,10 @@ static qint Parse_EvaluateTokens(source_t *source, token_t *tokens, qint *intval
       //remove the second value if not question mark operator
       if (o->operator != P_QUESTIONMARK) v = v->next;
       //
-      if (v)
-      {
-        if (v->prev) v->prev->next = v->next;
-        else firstvalue = v->next;
-        if (v->next) v->next->prev = v->prev;
-      }
+      if (v->prev) v->prev->next = v->next;
+      else firstvalue = v->next;
+      if (v->next) v->next->prev = v->prev;
+      else lastvalue = v->prev;
       //Z_Free(v);
       FreeValue(v);
     }
@@ -2333,6 +2303,7 @@ static qint Parse_EvaluateTokens(source_t *source, token_t *tokens, qint *intval
     if (o->prev) o->prev->next = o->next;
     else firstoperator = o->next;
     if (o->next) o->next->prev = o->prev;
+    else lastoperator = o->prev;
     //Z_Free(o);
     FreeOperator(o);
   }
@@ -2364,7 +2335,8 @@ static qint Parse_EvaluateTokens(source_t *source, token_t *tokens, qint *intval
 Parse_Evaluate
 ===============
 */
-static qint Parse_Evaluate(source_t *source, qint *intvalue, float *floatvalue, qint integer)
+static qint Parse_Evaluate(source_t *source, signed long qint *intvalue,
+                        double *floatvalue, qint integer)
 {
   token_t token, *firsttoken, *lasttoken;
   token_t *t, *nexttoken;
@@ -2448,7 +2420,8 @@ static qint Parse_Evaluate(source_t *source, qint *intvalue, float *floatvalue, 
 Parse_DollarEvaluate
 ===============
 */
-static qint Parse_DollarEvaluate(source_t *source, qint *intvalue, float *floatvalue, qint integer)
+static qint Parse_DollarEvaluate(source_t *source, signed long qint *intvalue,
+                        double *floatvalue, qint integer)
 {
   qint indent, defined = qfalse;
   token_t token, *firsttoken, *lasttoken;
@@ -2705,7 +2678,7 @@ Parse_Directive_elif
 */
 static qint Parse_Directive_elif(source_t *source)
 {
-  qint value;
+  signed long qint value;
   qint type, skip;
 
   Parse_PopIndent(source, &type, &skip);
@@ -2727,7 +2700,7 @@ Parse_Directive_if
 */
 static qint Parse_Directive_if(source_t *source)
 {
-  qint value;
+  signed long qint value;
   qint skip;
 
   if (!Parse_Evaluate(source, &value, NULL, qtrue)) return qfalse;
@@ -2802,7 +2775,7 @@ Parse_Directive_eval
 */
 static qint Parse_Directive_eval(source_t *source)
 {
-  qint value;
+  signed long qint value;
   token_t token;
 
   if (!Parse_Evaluate(source, &value, NULL, qtrue)) return qfalse;
@@ -2811,7 +2784,7 @@ static qint Parse_Directive_eval(source_t *source)
   token.whitespace_p = source->scriptstack->script_p;
   token.endwhitespace_p = source->scriptstack->script_p;
   token.linescrossed = 0;
-  sprintf(token.string, "%d", abs(value));
+  sprintf(token.string, "%ld", labs(value));
   token.type = TT_NUMBER;
   token.subtype = TT_INTEGER|TT_LONG|TT_DECIMAL;
   Parse_UnreadSourceToken(source, &token);
@@ -2826,7 +2799,7 @@ Parse_Directive_evalfloat
 */
 static qint Parse_Directive_evalfloat(source_t *source)
 {
-  float value;
+  double value;
   token_t token;
 
   if (!Parse_Evaluate(source, NULL, &value, qfalse)) return qfalse;
@@ -2849,7 +2822,7 @@ Parse_DollarDirective_evalint
 */
 static qint Parse_DollarDirective_evalint(source_t *source)
 {
-  qint value;
+  signed long qint value;
   token_t token;
 
   if (!Parse_DollarEvaluate(source, &value, NULL, qtrue)) return qfalse;
@@ -2858,7 +2831,7 @@ static qint Parse_DollarDirective_evalint(source_t *source)
   token.whitespace_p = source->scriptstack->script_p;
   token.endwhitespace_p = source->scriptstack->script_p;
   token.linescrossed = 0;
-  sprintf(token.string, "%d", abs(value));
+  sprintf(token.string, "%ld", labs(value));
   token.type = TT_NUMBER;
   token.subtype = TT_INTEGER|TT_LONG|TT_DECIMAL;
   token.intvalue = value;
@@ -2875,7 +2848,7 @@ Parse_DollarDirective_evalfloat
 */
 static qint Parse_DollarDirective_evalfloat(source_t *source)
 {
-  float value;
+  double value;
   token_t token;
 
   if (!Parse_DollarEvaluate(source, NULL, &value, qfalse)) return qfalse;
@@ -3035,7 +3008,7 @@ static qint Parse_Directive_endif(source_t *source)
 Parse_CheckTokenString
 ===============
 */
-static qint Parse_CheckTokenString(source_t *source, const qchar *string)
+static qint Parse_CheckTokenString(source_t *source, qchar *string)
 {
   token_t tok;
 
@@ -3083,11 +3056,13 @@ static qint Parse_Directive_define(source_t *source)
     //unread the define name before executing the #undef directive
     Parse_UnreadSourceToken(source, &token);
     if (!Parse_Directive_undef(source)) return qfalse;
+    //if the define was not removed (define->flags & DEFINE_FIXED)
+    define = Parse_FindHashedDefine(source->definehash, token.string);
   }
   //allocate define
-  define = (define_t *)Z_Malloc(sizeof(define_t));
+  define = (define_t *) Z_Malloc(sizeof(define_t) + strlen(token.string) + 1);
   Com_Memset(define, 0, sizeof(define_t));
-  define->name = (qchar *)Z_Malloc(strlen(token.string) + 1);
+  define->name = (qchar *) define + sizeof(define_t);
   strcpy(define->name, token.string);
   //add the define to the source
   Parse_AddDefineToHash(define, source->definehash);
@@ -3245,6 +3220,141 @@ static void Parse_UnreadToken(source_t *source, token_t *token)
 
 /*
 ===============
+Parse_ReadEnumeration
+
+It is assumed that the 'enum' token has already been consumed
+This is fairly basic: it doesn't catch some fairly obvious errors like nested
+enums, and enumerated names conflict with #define parameters
+===============
+*/
+static qbool Parse_ReadEnumeration( source_t *source )
+{
+  token_t newtoken;
+  qint value;
+
+  if( !Parse_ReadToken( source, &newtoken ) )
+    return qfalse;
+
+  if( newtoken.type != TT_PUNCTUATION || newtoken.subtype != P_BRACEOPEN )
+  {
+    Parse_SourceError( source, "Found %s when expecting {\n",
+                       newtoken.string );
+    return qfalse;
+  }
+
+  for( value = 0;; value++ )
+  {
+    token_t name;
+
+    // read the name
+    if( !Parse_ReadToken( source, &name ) )
+      break;
+
+    // it's ok for the enum to end immediately
+    if( name.type == TT_PUNCTUATION && name.subtype == P_BRACECLOSE )
+    {
+      if( !Parse_ReadToken( source, &name ) )
+        break;
+
+      // ignore trailing semicolon
+      if( name.type != TT_PUNCTUATION || name.subtype != P_SEMICOLON )
+        Parse_UnreadToken( source, &name );
+
+      return qtrue;
+    }
+
+    // ... but not for it to do anything else
+    if( name.type != TT_NAME )
+    {
+      Parse_SourceError( source, "Found %s when expecting identifier\n",
+                         name.string );
+      return qfalse;
+    }
+
+    if( !Parse_ReadToken( source, &newtoken ) )
+      break;
+
+    if( newtoken.type != TT_PUNCTUATION )
+    {
+      Parse_SourceError( source, "Found %s when expecting , or = or }\n",
+                         newtoken.string );
+      return qfalse;
+    }
+
+    if( newtoken.subtype == P_ASSIGN )
+    {
+      qint neg = 1;
+
+      if( !Parse_ReadToken( source, &newtoken ) )
+        break;
+
+      // Parse_ReadToken doesn't seem to read negative numbers, so we do it
+      // ourselves
+      if( newtoken.type == TT_PUNCTUATION && newtoken.subtype == P_SUB )
+      {
+        neg = -1;
+
+        // the next token should be the number
+        if( !Parse_ReadToken( source, &newtoken ) )
+          break;
+      }
+
+      if( newtoken.type != TT_NUMBER || !( newtoken.subtype & TT_INTEGER ) )
+      {
+        Parse_SourceError( source, "Found %s when expecting integer\n",
+                           newtoken.string );
+        return qfalse;
+      }
+
+      // this is somewhat silly, but cheap to check
+      if( neg == -1 && ( newtoken.subtype & TT_UNSIGNED ) )
+      {
+        Parse_SourceWarning( source, "Value in enumeration is negative and "
+                                     "unsigned\n" );
+      }
+
+      // set the new define value
+      value = newtoken.intvalue * neg;
+
+      if( !Parse_ReadToken( source, &newtoken ) )
+        break;
+    }
+
+    if( newtoken.type != TT_PUNCTUATION || ( newtoken.subtype != P_COMMA &&
+        newtoken.subtype != P_BRACECLOSE ) )
+    {
+      Parse_SourceError( source, "Found %s when expecting , or }\n",
+                         newtoken.string );
+      return qfalse;
+    }
+
+    if( !Parse_AddDefineToSourceFromString( source, va( "%s %d\n", name.string,
+                                                        value ) ) )
+    {
+      Parse_SourceWarning( source, "Couldn't add define to source: %s = %d\n",
+                           name.string, value );
+      return qfalse;
+    }
+
+    if( newtoken.subtype == P_BRACECLOSE )
+    {
+      if( !Parse_ReadToken( source, &name ) )
+        break;
+
+      // ignore trailing semicolon
+      if( name.type != TT_PUNCTUATION || name.subtype != P_SEMICOLON )
+        Parse_UnreadToken( source, &name );
+
+      return qtrue;
+    }
+  }
+
+  // got here if a ReadToken returned false
+  return qfalse;
+}
+
+/*
+===============
 Parse_ReadToken
 ===============
 */
@@ -3272,6 +3382,12 @@ static qint Parse_ReadToken(source_t *source, token_t *token)
         continue;
       }
     }
+    if( token->type == TT_NAME && !Q_stricmp( token->string, "enum" ) )
+    {
+      if( !Parse_ReadEnumeration( source ) )
+        return qfalse;
+      continue;
+    }
     // recursively concatenate strings that are behind each other still resolving defines
     if (token->type == TT_STRING)
     {
@@ -3283,10 +3399,10 @@ static qint Parse_ReadToken(source_t *source, token_t *token)
           token->string[strlen(token->string)-1] = '\0';
           if (strlen(token->string) + strlen(newtoken.string+1) + 1 >= MAX_TOKEN_CHARS)
           {
-            Parse_SourceError(source, "string longer than MAX_TOKEN_CHARS %d", MAX_TOKEN_CHARS);
+            Parse_SourceError(source, "string longer than MAX_TOKEN_CHARS %d\n", MAX_TOKEN_CHARS);
             return qfalse;
           }
-          strcat(token->string, newtoken.string + 1);
+          strcat(token->string, newtoken.string+1);
         }
         else
         {
@@ -3321,7 +3437,7 @@ static qint Parse_ReadToken(source_t *source, token_t *token)
 Parse_DefineFromString
 ===============
 */
-static define_t *Parse_DefineFromString(const qchar *string)
+static define_t *Parse_DefineFromString(qchar *string)
 {
   script_t *script;
   source_t src;
@@ -3332,7 +3448,7 @@ static define_t *Parse_DefineFromString(const qchar *string)
   script = Parse_LoadScriptMemory(string, strlen(string), "*extern");
   //create a new source
   Com_Memset(&src, 0, sizeof(source_t));
-  Q_strncpyz(src.filename, "*extern", sizeof(src.filename));
+  strncpy(src.filename, "*extern", MAX_QPATH);
   src.scriptstack = script;
   src.definehash = Z_Malloc(DEFINEHASHSIZE * sizeof(define_t *));
   Com_Memset( src.definehash, 0, DEFINEHASHSIZE * sizeof(define_t *));
@@ -3357,12 +3473,25 @@ static define_t *Parse_DefineFromString(const qchar *string)
   Z_Free(src.definehash);
   //
   Parse_FreeScript(script);
-  //if the define was created successfully
+  //if the define was created succesfully
   if (res > 0) return def;
-  //free the define if created
+  //free the define is created
   if (src.defines) Parse_FreeDefine(def);
   //
   return NULL;
+}
+
+/*
+===============
+Parse_AddDefineToSourceFromString
+===============
+*/
+static qbool Parse_AddDefineToSourceFromString( source_t *source,
+                                                   const qchar *string )
+{
+  Parse_PushScript( source, Parse_LoadScriptMemory( string, strlen( string ),
+                                                    "*extern" ) );
+  return Parse_Directive_define( source );
 }
 
 /*
@@ -3372,7 +3501,7 @@ Parse_AddGlobalDefine
 add a globals define that will be added to all opened sources
 ===============
 */
-qint Parse_AddGlobalDefine(const qchar *string)
+qint Parse_AddGlobalDefine(qchar *string)
 {
   define_t *define;
 
@@ -3388,12 +3517,12 @@ qint Parse_AddGlobalDefine(const qchar *string)
 Parse_CopyDefine
 ===============
 */
-static define_t *Parse_CopyDefine(source_t *source, const define_t *define)
+static define_t *Parse_CopyDefine(source_t *source, define_t *define)
 {
   define_t *newdefine;
   token_t *token, *newtoken, *lasttoken;
 
-  newdefine = (define_t *)Z_Malloc(sizeof(define_t));
+  newdefine = (define_t *) Z_Malloc(sizeof(define_t) + strlen(define->name) + 1);
   //copy the define name
   newdefine->name = (qchar *) newdefine + sizeof(define_t);
   strcpy(newdefine->name, define->name);
@@ -3433,8 +3562,7 @@ Parse_AddGlobalDefinesToSource
 */
 static void Parse_AddGlobalDefinesToSource(source_t *source)
 {
-  const define_t *define;
-  define_t *newdefine;
+  define_t *define, *newdefine;
 
   for (define = globaldefines; define; define = define->next)
   {
@@ -3458,8 +3586,8 @@ static source_t *Parse_LoadSourceFile(const qchar *filename)
 
   script->next = NULL;
 
-  source = (source_t *) Z_Malloc(sizeof(*source));
-  Com_Memset(source, 0, sizeof(*source));
+  source = (source_t *) Z_Malloc(sizeof(source_t));
+  Com_Memset(source, 0, sizeof(source_t));
 
   Q_strncpyz(source->filename, filename, sizeof(source->filename));
   source->scriptstack = script;
@@ -3545,7 +3673,6 @@ qint Parse_LoadSourceHandle(const qchar *filename)
   }
   if (i >= MAX_SOURCEFILES)
     return 0;
-  Parse_SetBaseFolder("");
   source = Parse_LoadSourceFile(filename);
   if (!source)
     return 0;